Changed file extensions, fixed issues and formatted files
